Restructured Options for Select Fields
Instead of a hash (which may get reordered when being stored as JSON) it now uses an array of OptionValues.

These will be stored in the correct order, so can be displayed consistently.

In addition, it allows us to update the scores and colours for each option as well.

There is a conversion routine built-in - so previously stored fields, with options hashes are converted to OptionValue arrays